% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{get_domains}
\alias{get_domains}
\title{Fetches domain names for searched universities}
\usage{
get_domains(name, country)
}
\arguments{
\item{name}{A one-element character string that contains all or part of a university name (e.g., "harvard", "middle georgia").}

\item{country}{A one-element character string that contains a country name (e.g., "turkey", "canada"). This is an optional element.}
}
\value{
An object that contains domain names for searched universities.
}
\description{
Returns an object that contains domain names for searched universities.
}
\examples{
\dontrun{
school <- c("virginia")
place <- c("canada")

y <- get_domains(school, place)
y
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
